be/requirements.txt
pyrealsense2
numpy
opencv-python-headless
PyQt5
PyOpenGl
Pillow
ultralytics
deep_sort_realtime
norfair